requirements.txt
# breadbox-core
fastapi[standard]~=0.115.12
aiofiles~=24.1.0
slowapi~=0.1.9
pydantic~=2.11.4
toml~=0.10.2
starlette~=0.46.2

# app
sqlmodel~=0.0.24
uvicorn~=0.34.2
argon2-cffi~=23.1.0